decouple user functions and user queries and seriously reduce code overhead
