+++ app/data/types.ts
@@ -3,7 +3,7 @@ import { IUser } from '../../models/types';
 
 export type UserDocument = IUser & Document
 export type UserQuery = UserDocument | null
-export enum UserQueries {
+export enum UserQueryTypes {
     getByUsername = 'getByUsername',
     getByEmail = 'getByEmail',
     getByEmailorUsername = 'getByEmailorUsername'
